Make on top of homepage MovieRama clickable
This is useful in case someone wants to
reset the UI to the root "/movies" page.

For example if I click on a username and then
press to sort the results of the user's movies
by likes, then I have no convenient way to go back
to the "/movies" page with all the users posts.

Clicking the MovieRama title on top is an
intuitive way to return to the base of the
website. i.e. The homepage.