Add error_go_back.html
Create this html template
and use it whenever an 40X error
occurs.

e.g. A user tries to login with
incorrect credentials:
In the pastit would be returned a
simple json and the corresponding
status code and error message.

Now this html page will be displayed
so that it will be more user friendly.